code/health-agent/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
httpx==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
# Optional (for production): gunicorn, prometheus-client, etc.
